Update Docker environment specs (#2060)
* Refresh Docker conda environment specs

* Bump min hdf5 version in soil moisture environment

* Update tools/imagesets/oracle8conda/distrib_nisar/requirements.soil-moisture.yml

Co-authored-by: Xiaodong Huang <Xiaodong.Huang@jpl.nasa.gov>

* Allow gdal>=3.9.2; Add libgdal-hdf5 & libgdal-netcdf

* Fix metadata_cubes test with newer gdal versions

The test was failing due to trying to read from a raster that had not
yet been flushed. The same dataset was opened for writing and for
reading simultaneously, which is not safe in general since changes
written to the raster may be cached until the raster object is
destroyed.

The fix is to ensure that the original raster object goes out of scope
(so that changes are guaranteed to be flushed) before re-opening it for
reading. In other cases, we instead simply read from the original raster
object to instead of re-opening it for reading.

---------

Co-authored-by: Xiaodong Huang <Xiaodong.Huang@jpl.nasa.gov>